finished working  on heuristic

+++ src/isolationGame/EvalFunc1.java
@@ -3,7 +3,7 @@ package isolationGame;
 public class EvalFunc1 extends Heuristic {
 
 	@Override
-	public int evalFunc(Board board) {
+	public double evalFunc(Board board) {
 	
 		int funcVal = board.getMoves(true) - board.getMoves(false);
