+++ src/isolationGame/Heuristic.java
@@ -1,6 +1,6 @@
 package isolationGame;
 
 abstract public class Heuristic {	
-	abstract public int evalFunc(Board board);
+	abstract public double evalFunc(Board board);
 		
 }
